+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Fancy Git Changelog
 
+### v6.1.6
+
+- Improve the way fancygit looks for its configurations.
+- Add the `fancygit --config-list` command to show the current FancyGit configuration.
+- Add the `fancygit --config-reset` command to reset FancyGit configuration.
+- Add the `fancygit --full-path-enable` command for making FancyGit to show the full path of the working directory.
+- Add the `fancygit --full-path-disable` command for making FancyGit to show only working directory, not the full path.
+
+
 ### v6.0.5
 
 - Add new icon when branch exists only locally. (Only colored themes)